moved piped input scanner to model #538
